requirements.txt
# OCR / PDF
pymupdf>=1.24.0
ollama>=0.4.0

# Web app
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
websockets>=12.0

# Database
psycopg2-binary>=2.9.9
sshtunnel>=0.4.0

# Config
python-dotenv>=1.0.0
